Initial Release of TextTemplateTransformationEngine
This drop-in function provides a simple template based text generation
engine.  It basically allows code to be encapsulated in special tags
that can manipulate or insert text. If you have worked with classic ASP,
JSP, PHP or T4 templates you're probably familiar with text template
transformations. In classic ASP it was done like <body>Current time:<br
/><%Response.Write Now()%></body>. In T4 you can do something like
Current time:<<#= DateTime.Now #>. And finally in this project, it can
be done like Current time: [=DateTime.Now ]] or Current time:
/*=DateTime.Now :*/.

Basically text transformations, also called dynamic text, allows the use
of programming methods to modify text. Common uses might be for
repeating sections of text, filling in fields on an ASP page, showing a
username or account code in an email, or to write 1 to 1000 on a
webpage.

This project is somewhat similar to Microsoft's T4 but much simpler. It
similar because it uses encapsulated C# code to format text. Microsoft
Visual Studio's T4 is much more powerful and this project is not meant
to be replacement...at least not inside of Visual Studio!  There are a
couple of issues when using T4 templates in a 3rd party applications.
The foremost is the licensing. The T4 DLL is not redistributable. There
are funky ways around this by installing some MS packages that have the
dll or installing visual studio express but that is messy.  This project
by contrast is not even a DLL, it is a simple drop in function.  Another
issue is T4 does not mesh well with many syntax highlighting and code
completion projects. To work around this I created \*: code-here :*\
like commands in comments.